feat(migrate): match install's progress bar TUI in serial and parallel
`mt migrate` and `mt migrate --parallel` now render per-keg progress
bars with the same cyan-filled / dim-empty look and the dim
`(MB | rate | ETA)` detail block as `mt install`'s download phase, and
finish each keg with a green ✓. The closing summary tightens to
left-aligned counts under `✓ Migration completed.` and adds a
copy-pasteable dim `brew uninstall <names>` hint so a successful
migration converges with a clean Homebrew-side teardown.

Non-TTY runs keep the existing per-keg log lines; bars are
visual-only.

+++ src/cli/install/download.zig
@@ -55,7 +55,7 @@ pub const DownloadJob = struct {
     /// Alignment width for progress bar labels (max name length across all jobs).
     label_width: u8,
     /// Line index within the multi-progress group.
-    line_index: u8,
+    line_index: u16,
     /// Shared multi-progress state for coordinated rendering.
     multi: ?*progress_mod.MultiProgress,
     /// Progress bar owned by the main thread. Created before workers are spawned
